// src/types.ts
// export enum HealthCheckRating {
//   "Healthy" = 0,
//   "LowRisk" = 1,
//   "HighRisk" = 2,
//   "CriticalRisk" = 3
// }

// export interface Diagnosis {
//   code: string;
//   name: string;
//   latin?: string;
// }

// interface BaseEntry {
//   id: string;
//   description: string;
//   date: string;
//   specialist: string;
//   diagnosisCodes?: Array<Diagnosis["code"]>;
// }

// // export enum EntryType {
// //     HealthCheck = 'HealthCheck',
// //     Hospital = 'Hospital',
// //     OccupationalHealthcare = 'OccupationalHealthcare'
// // }

// export interface HealthCheckEntry extends BaseEntry {
//   type: "HealthCheck";
//   healthCheckRating: HealthCheckRating;
// }

// export interface HospitalEntry extends BaseEntry {
//   type: "Hospital";
//   discharge: {
//     date: string;
//     criteria: string;
//   };
// }

// export interface OccupationalHealthcareEntry extends BaseEntry {
//   type: "OccupationalHealthcare";
//   employerName: string;
//   sickLeave?: {
//     startDate: string;
//     endDate: string;
//   };
// }

// export enum Gender {
//   Male = 'male',
//   Female = 'female',
//   Other = 'other'
// }


// export interface Patient {
//   id: string;
//   name: string;
//   dateOfBirth?: string;
//   ssn?: string;
//   gender: Gender;
//   occupation: string;
//   entries?: Entry[];
// }

// export type NewPatient = Omit<Patient, 'id'>;
// export type PublicPatient = Omit<Patient, 'ssn' | 'entries'>;
// export type Entry =
//   | HospitalEntry
//   | OccupationalHealthcareEntry
//   | HealthCheckEntry;



export enum HealthCheckRating {
  "Healthy" = 0,
  "LowRisk" = 1,
  "HighRisk" = 2,
  "CriticalRisk" = 3
}

export interface Diagnosis {
  code: string;
  name: string;
  latin?: string;
}

interface BaseEntry {
  id: string;
  description: string;
  date: string;
  specialist: string;
  diagnosisCodes?: Array<Diagnosis["code"]>;
}

export enum EntryType {
  HealthCheck = 'HealthCheck',
  Hospital = 'Hospital',
  OccupationalHealthcare = 'OccupationalHealthcare'
}

export interface HealthCheckEntry extends BaseEntry {
  type: "HealthCheck";
  healthCheckRating: HealthCheckRating;
}

export interface HospitalEntry extends BaseEntry {
  type: "Hospital";
  discharge: {
    date: string;
    criteria: string;
  };
}

export interface OccupationalHealthcareEntry extends BaseEntry {
  type: "OccupationalHealthcare";
  employerName: string;
  sickLeave?: {
    startDate: string;
    endDate: string;
  };
}

export enum Gender {
  Male = 'male',
  Female = 'female',
  Other = 'other'
}


export interface Patient {
  id: string;
  name: string;
  dateOfBirth: string;
  ssn: string;
  gender: Gender;
  occupation: string;
  entries: Entry[];
}

export type NewPatient = Omit<Patient, 'id'>;
export type PublicPatient = Omit<Patient, 'ssn' | 'entries'>;
export type Entry =
  | HospitalEntry
  | OccupationalHealthcareEntry
  | HealthCheckEntry;


export type NewHospitalEntry = Omit<HospitalEntry, 'id'>;
export type NewOccupationalHealthcareEntry = Omit<OccupationalHealthcareEntry, 'id'>;
export type NewHealthCheckEntry = Omit<HealthCheckEntry, 'id'>;
export type NewEntry =
  | NewHospitalEntry
  | NewOccupationalHealthcareEntry
  | NewHealthCheckEntry;